Menu Items upgraded
