+++ app/src/main/java/read/code/yourreader/Fragments/HomeFragment.kt
@@ -1,0 +1,29 @@
+package read.code.yourreader.Fragments
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import read.code.yourreader.R
+
+
+class HomeFragment : Fragment() {
+
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        var view= inflater.inflate(R.layout.fragment_home, container, false)
+
+        return view
+    }
+
+}
