+++ app/src/main/res/layout/activity_login.xml
@@ -2,11 +2,10 @@
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:theme="@style/Theme.AppTheme"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
 
-    tools:context=".Login.LoginActivity">
+    tools:context=".Auth.LoginActivity">
 
 
     <RelativeLayout
