+++ app/src/main/res/menu/menu1.xml
@@ -3,6 +3,11 @@
 
     <group>
 
+        <item
+            android:id="@+id/home_menu"
+            android:icon="@drawable/ic_home"
+            android:title="@string/home" />
+
 
         <item
             android:id="@+id/reading_now"
@@ -12,30 +17,24 @@
             android:id="@+id/books"
             android:icon="@drawable/ic_books"
             android:title="@string/books" />
-        <item
-            android:id="@+id/docu"
-            android:icon="@drawable/ic_books"
-            android:title="@string/documents" />
+
         <item
             android:id="@+id/menu_haveread"
             android:icon="@drawable/ic_baseline_done_24"
             android:title="@string/have_read" />
 
-        <!--        What is use of following?
-                    following tabs seems confusing-->
-        <!--        <item-->
-        <!--            android:id="@+id/formats_menu"-->
-        <!--            android:icon="@drawable/ic_format"-->
-        <!--            android:title="@string/formats" />-->
 
-        <!--        <item-->
-        <!--            android:id="@+id/menu_Folders"-->
-        <!--            android:icon="@drawable/ic_folders"-->
-        <!--            android:title="@string/folders" />-->
-        <!--        <item-->
-        <!--            android:id="@+id/menu_Downlaods"-->
-        <!--            android:icon="@drawable/ic_download"-->
-        <!--            android:title="@string/downlaods" />-->
+        <item
+            android:id="@+id/menu_Downloads"
+            android:icon="@drawable/ic_download"
+            android:title="@string/downlaods" />
+
+        <item
+            android:id="@+id/menu_use"
+            android:icon="@drawable/ic_use"
+            android:title="@string/how_to_use" />
+
+
     </group>
     <group>
         <item
