+++ app/src/main/res/values/strings.xml
@@ -20,16 +20,18 @@
     <string name="or_continue_with">Or Continue with</string>
     <string name="signout">signOut</string>
     <string name="reading_now">Reading Now</string>
-    <string name="books">Books </string>
+    <string name="books">Books and Documents </string>
     <string name="documents">Documents</string>
     <string name="have_read">Done Reading</string>
     <string name="formats">Formats</string>
     <string name="folders">Folders</string>
-    <string name="downlaods">Downlaods</string>
+    <string name="downlaods">Downloads</string>
     <string name="settings">Settings</string>
     <string name="send_feedback">Send Feedback</string>
     <string name="close">Close</string>
     <string name="open">Open</string>
     <string name="todo">TODO</string>
     <string name="hello_blank_fragment">Hello blank fragment</string>
+    <string name="how_to_use">How to use</string>
+    <string name="home">Home</string>
 </resources>
